requirements (1).txt
# Django Conversation Backend Requirements
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
celery==5.3.4
redis==5.0.1
django-extensions==3.2.3
